Fixed an issue where scheduled messages were not being sent (#94)
Add/remove rule scheduler depending on the status of the parent rule (#98)
Added a phpunit.xml.dist file.
Added a bootstrap.php file for unit tests.
Cron will now de-activate a rule (message) once it has been added to the queue.
Fixed an issue where the ScheduledRuleProcessor worker would the incorrect entity type.
Renamed test trait createRules to createUserRoleRules.
Added tests for rule schedule entity creation and deletion depending on parent rule status.
Injected event manager service into ScheduledRuleProcessor.
Added extra guards in ScheduledRuleProcessor in case the rule schedule entity was deleted.
Added guards to template collection action to make sure it has correct context.

+++ src/Plugin/QueueWorker/ScheduledRuleProcessor.php
@@ -7,7 +7,10 @@
 
 namespace Drupal\rng\Plugin\QueueWorker;
 
+use Symfony\Component\DependencyInjection\ContainerInterface;
 use Drupal\Core\Queue\QueueWorkerBase;
+use Drupal\Core\Plugin\ContainerFactoryPluginInterface;
+use Drupal\rng\EventManagerInterface;
 use Drupal\rng\Entity\RuleSchedule;
 
 /**
@@ -19,26 +22,74 @@ use Drupal\rng\Entity\RuleSchedule;
  *   cron = {"time" = 60}
  * )
  */
-class ScheduledRuleProcessor extends QueueWorkerBase {
+class ScheduledRuleProcessor extends QueueWorkerBase implements ContainerFactoryPluginInterface {
+
+  /**
+   * The RNG event manager.
+   *
+   * @var \Drupal\rng\EventManagerInterface
+   */
+  protected $eventManager;
+
+  /**
+   * Constructs a ScheduledRuleProcessor object.
+   *
+   * @param array $configuration
+   *   A configuration array containing information about the plugin instance.
+   * @param string $plugin_id
+   *   The plugin_id for the plugin instance.
+   * @param mixed $plugin_definition
+   *   The plugin implementation definition.
+   * @param \Drupal\rng\EventManagerInterface $event_manager
+   *   The RNG event manager.
+   */
+  public function __construct(array $configuration, $plugin_id, array $plugin_definition, EventManagerInterface $event_manager) {
+    parent::__construct($configuration, $plugin_id, $plugin_definition);
+    $this->eventManager = $event_manager;
+  }
+
+  /**
+   * {@inheritdoc}
+   */
+  public static function create(ContainerInterface $container, array $configuration, $plugin_id, $plugin_definition) {
+    return new static(
+      $configuration,
+      $plugin_id,
+      $plugin_definition,
+      $container->get('rng.event_manager')
+    );
+  }
 
   /**
    * {@inheritdoc}
    *
-   * @param $data
-   *   - rule_component_id: integer: ID of a rule component entity.
+   * @param integer $data['rule_scheduler_id']
+   *   ID of a rule component entity.
    */
   public function processItem($data) {
-    /** @var RuleSchedule $rule_schedule */
-    $rule_schedule = RuleSchedule::load($data['rule_component_id']);
+    if (!isset($data['rule_scheduler_id'])) {
+      return;
+    }
+
+    if (!$rule_schedule = RuleSchedule::load($data['rule_scheduler_id'])) {
+      return;
+    }
+
     $rule_schedule->incrementAttempts();
     $rule_schedule->save();
-    if ($component = $rule_schedule->getComponent()) {
-      $rule = $component->getRule();
-      $event_meta = \Drupal::service('rng.event_manager')
-        ->getMeta($rule->getEvent());
-      $event_meta->trigger($rule->getTriggerID(), [
+
+    if (($component = $rule_schedule->getComponent()) && ($rule = $component->getRule())) {
+      $event = $rule->getEvent();
+      $event_meta = $this->eventManager->getMeta($event);
+      $context = [
+        'event' => $event,
         'registrations' => $event_meta->getRegistrations()
-      ]);
+      ];
+
+      foreach ($rule->getActions() as $action) {
+        $action->execute($context);
+      }
+
       $rule_schedule->delete();
     }
   }
